--- components/common/AddressInput/index.js
@@ -1,3 +1,0 @@
-import AddressInput from './AddressInput';
-
-export { AddressInput };
