+++ components/common/AddressPostcode/index.js
@@ -1,0 +1,3 @@
+import AddressPostcode from './AddressPostcode';
+
+export { AddressPostcode };
