+++ components/common/index.js
@@ -1,4 +1,4 @@
-import { AddressInput } from './AddressInput';
+import { AddressPostcode } from './AddressPostcode';
 import { ButtonL, ButtonM, ButtonS } from './Button';
 import { CheckBox } from './CheckBox';
 import { Dropdown } from './Dropdown';
@@ -15,7 +15,7 @@ import { Title } from './Title';
 import { Toast } from './Toast';
 
 export {
-  AddressInput,
+  AddressPostcode,
   ButtonL,
   ButtonM,
   ButtonS,
