+++ package.json
@@ -16,6 +16,7 @@
     "next": "^14.2.17",
     "next-auth": "^5.0.0-beta.25",
     "react": "^18.3.1",
+    "react-daum-postcode": "^3.1.3",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
     "react-infinite-scroll-component": "^6.1.0",
